requirements.txt
astroid==3.1.0
coverage==7.4.4
dill==0.3.8
exceptiongroup==1.2.0
Faker==24.2.0
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
numpy==1.24.4
packaging==24.0
pandas==2.0.3
platformdirs==4.2.0
pluggy==1.4.0
pylint==3.1.0
pytest==8.1.1
pytest-cov==4.1.0
pytest-pylint==0.21.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
tomli==2.0.1
tomlkit==0.12.4
typing-extensions==4.10.0
tzdata==2024.1
